Move patch_fixture() to pytest_localstack
