+++ tests/functional/test_patch_fixture/test_README.py
@@ -6,7 +6,7 @@ import pytest
 import pytest_localstack
 
 
-patch_s3 = pytest_localstack.patch_fixture(services=["s3"])  # type: ignore
+patch_s3 = pytest_localstack.patch_fixture(services=["s3"])
 
 
 @pytest.mark.usefixtures("patch_s3")
