+++ tests/functional/test_patch_fixture/test_services_available.py
@@ -5,7 +5,7 @@ import botocore
 import pytest_localstack
 
 
-localstack = pytest_localstack.patch_fixture(scope="module", autouse=True)  # type: ignore
+localstack = pytest_localstack.patch_fixture(scope="module", autouse=True)
 
 
 def _assert_key_isinstance(result, key, type):
